Move sanitization directly to the model-checking library
