+++ classifier/Cargo.toml
@@ -7,14 +7,10 @@ license = "MIT"
 edition = "2021"
 rust-version = "1.60"
 
-[[bin]]
-name = "load-classifier-output"
-path = "src/bin/load_classifier_output.rs"
-
 [dependencies]
-biodivine-lib-bdd = ">=0.4.2, <1.0.0"
-biodivine-lib-param-bn = ">=0.4.3, <1.0.0"
-biodivine-hctl-model-checker = ">=0.1.2, <1.0.0"
+biodivine-lib-bdd = ">=0.5.1, <1.0.0"
+biodivine-lib-param-bn = ">=0.4.5, <1.0.0"
+biodivine-hctl-model-checker = ">=0.1.6, <1.0.0"
 clap = { version = "4.1.1", features = ["derive"] }
 zip = "0.6.3"
